backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.21.0
google-cloud-aiplatform==1.68.0
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
